requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
sqlalchemy==2.0.20
alembic==1.11.3
pydantic[email]==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
psycopg2-binary
python-dotenv==1.0.0
tenacity==8.2.3
emails==0.6
jinja2==3.1.2
celery==5.3.1
redis==4.6.0
pydantic_settings
asyncpg
watchdog